Add basic functionality for filtering the results
Within the Market class is a filterHash, which stores the fields that are not 0 (within a binary format).
Within the Hauptfenster_Controller there happens some binary magic.
This binary magic returns 0, if the filter is ok.
At the moment nothing is filtered.
The first 0 of the line "if((0 & (~markt.getFilterHash())) == 0) {" must be the filter in binary format.